% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_analysis.R
\name{comp_res}
\alias{comp_res}
\title{Function to plot deconvolution and reference matrices.}
\usage{
comp_res(res = res_, ref = ref_, pipe = TRUE)
}
\arguments{
\item{res}{res, a list(A_matrix, T_matrix)}

\item{ref, }{a list(A_matrix, T_matrix)}
}
\value{
With pipe=TRUE (default) this function return res, a list(A_matrix, T_matrix)
}
\description{
This function compares a deconvolution result with its reference
}
\details{
#'pipes from ( ... \%>\% )  :  \link{run_deconv}nts <- sweep(mix_matrix*1e6 , 2, colSums(mix_matrix), FUN="/")

pipes to ( \%>\% ... ) :  \link{score_res}

with pipe=TRUE, returns a list(A_matrix, T_matrix)
}
