% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_analysis.R
\name{merge_hc}
\alias{merge_hc}
\title{Title}
\usage{
merge_hc(A, k, T = NULL, plot = NULL, unknown_est = NULL)
}
\arguments{
\item{A}{}

\item{k}{}

\item{plot}{}

\item{unknown_est}{}
}
\value{

}
\description{
Title
}
