% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deconvolution.R
\name{run_deconv}
\alias{run_deconv}
\title{Function to run deconvolution methods.}
\usage{
run_deconv(mix_matrix, k = 5, method)
}
\arguments{
\item{mix_matrix}{The convoluted pre-treated matrix patients*probes.}

\item{k}{The number of cell types to estimate}

\item{method}{The deconvolution method to be used}
}
\value{
This function return a list with the A and the T matrices computed.
}
\description{
This function run a deconvolution  methods and takes as input a pre-treated gene expression matrix
}
