% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_featsel.R
\name{run_featsel}
\alias{run_featsel}
\title{Function to run RNAseq feature selection}
\usage{
run_featsel(mix_matrix, method)
}
\arguments{
\item{mix_matrix}{The gene expression matrix samples \emph{x} genes.}

\item{method}{The feature selection method to be used}
}
\value{
This function return a matrix  samples \emph{x} genes
}
\description{
This function run a RNAseq feature selection methods and takes as input a gene expression matrix
}
\details{
available methods are cv and none

#'pipes from ( ... \%>\% )  : \link{data.matrix} , \link{run_norm} , \link{run_trans}

pipes to ( \%>\% ... ) : \link{run_deconv}

returns a matrix
}
