% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trans.R
\name{run_trans}
\alias{run_trans}
\title{Function to run RNAseq count transformation}
\usage{
run_trans(mix_matrix, method)
}
\arguments{
\item{mix_matrix}{The gene expression matrix samples \emph{x} genes.}

\item{method}{The transformation method to be used}
}
\value{
This function return a matrix  samples \emph{x} genes
}
\description{
This function run a RNAseq count transformation methods and takes as input a raw gene expression matrix
}
\details{
available methods are log2 pseudoLog (pseudolog)

log2 : f(x)=log2(x+1)

pseudoLog : f(x)=asinh(x)

pipes from ( ... \%>\% )  : \link{data.matrix} , \link{run_norm}

pipes to ( \%>\% ... ) : \link{run_deconv}

returns a matrix
}
