% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_analysis.R
\name{score_res}
\alias{score_res}
\title{Function to evaluate deconvolution and reference matrices.}
\usage{
score_res(res = res_, ref = ref_, pipe = TRUE)
}
\arguments{
\item{res}{res, a list(A_matrix, T_matrix)}

\item{ref, }{a list(A_matrix, T_matrix)}
}
\value{
With pipe=TRUE (default) this function return res, a list(A_matrix, T_matrix)
}
\description{
This function evaluates a deconvolution result with its reference
}
\details{
#'pipes from ( ... \%>\% )  :  \link{run_deconv}

pipes to ( \%>\% ... ) :  \link{comp_res}

with pipe=TRUE, returns a list(A_matrix, T_matrix)
}
