% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_norm.R
\name{run_norm}
\alias{run_norm}
\title{Function to run RNAseq count normalisation methods.}
\usage{
run_norm(mix_matrix, method, gene_length_bp = NULL)
}
\arguments{
\item{mix_matrix}{The raw gene count matrix samples \emph{x} genes.}

\item{method}{The normalization method to be used}

\item{meta}{The count matrix metadata}
}
\value{
This function return a matrix  samples \emph{x} genes
}
\description{
This function run a RNAseq count normalisation methods and takes as input a raw gene expression matrix
}
\details{
available methods are "DSEq2", "TPM", "RPM", "CPM", "edgeR"

TPM  and RKPM methods require gene length argument

#'pipes from ( ... \%>\% )  :  \link{data.matrix}

pipes to ( \%>\% ... ) :  \link{run_trans} , \link{run_deconv}

returns a matrix
}
